+++ spec/helpers/config_spec.rb
@@ -19,4 +19,21 @@ describe "Boot Strap Config" do
 
     expect(bootstrap_version).to be >= default_version
   end
+
+  it "doesn't contain unexpected entries" do
+    require 'toml-rb'
+    config = TomlRB.load_file("buildpack.toml")
+
+    urls = config["publish"]["Vendor"].map {|h| h["url"] if h["dir"] != "." }.compact
+    heroku_20 = urls.find_all {|url| url.include?("heroku-20") }
+    expect(heroku_20.length).to eq(1)
+
+    heroku_22 = urls.find_all {|url| url.include?("heroku-22") }
+    expect(heroku_22.length).to eq(1)
+
+    heroku_24 = urls.find_all {|url| url.include?("heroku-24") }
+    expect(heroku_24.length).to eq(1)
+
+    expect(urls.length).to eq(3)
+  end
 end